requirements.txt
docopt
python-can
mock